Move awaitilities struct to wait package
